# Fix load time calculation error in llama_bench