--- a/include/llama.h
+++ b/include/llama.h
@@ -414,7 +414,7 @@ extern "C" {
                              const char * path_model,
               struct llama_model_params   params);
 
-    LLAMA_API void llama_model_reset_time(struct llama_model * model);
+    LLAMA_API void llama_reset_model_time(struct llama_model * model);
 
     LLAMA_API void llama_free_model(struct llama_model * model);