--- a/src/llama.cpp
+++ b/src/llama.cpp
@@ -8809,6 +8809,11 @@ static bool llm_load_tensors(
     return true;
 }
 
+void llama_model_reset_time(llama_model * model) {
+    model->t_start_us = ggml_time_us();
+    model->t_load_us = ggml_time_us() - model->t_start_us;
+}
+
 // Returns 0 on success, -1 on error, and -2 on cancellation via llama_progress_callback
 static int llama_model_load(const std::string & fname, llama_model & model, llama_model_params & params) {
     model.t_start_us = ggml_time_us();