@slaren 2024-09-19
This is not a good solution, we should not add an API to workaround a bug. Instead, this needs to be fixed in llama.cpp so that the model loading times are not overwritten when creating additional contexts.