@slaren 2024-09-19
Alternatively, simply removing the update of the load time after the first evaluation would be enough. This was done to improve the accuracy of the load time when using mmap, since the model data might not have been read from this until it is used, but there are better ways to do that, and I am not sure that it is really that important.